+++ src/3D/L3DMesh.h
@@ -87,7 +87,7 @@ public:
 	bool LoadFromFile(const fs::path& path);
 	void LoadFromBuffer(const std::vector<uint8_t>& data);
 
-	[[nodiscard]] uint8_t GetNumSubMeshes() const { return _subMeshes.size(); }
+	[[nodiscard]] uint8_t GetNumSubMeshes() const { return static_cast<uint8_t>(_subMeshes.size()); }
 	[[nodiscard]] const std::vector<std::unique_ptr<L3DSubMesh>>& GetSubMeshes() const { return _subMeshes; }
 	[[nodiscard]] const std::unordered_map<SkinId, std::unique_ptr<graphics::Texture2D>>& GetSkins() const { return _skins; }
 	[[nodiscard]] const std::vector<uint32_t>& GetBoneParents() const { return _bonesParents; }
